When you see it, you won't be able to unsee it.
